agent/requirements.txt
flask
flask-cors
google-generativeai
sqlite-utils
python-dotenv
